Embed ChangeSet within Commit
Addresses concerns on #100.

This version still maintains two copies of a ChangeSet, one in
the Commit and one in the list of ChangeSets that survived the filter.

This is because the SCM interface doesn't have a way to pass in
(and save a reference to) a Changeset. Just the Id is passed at the moment.