Minimize time spent at max ChangeSet footprint
Problem:
We currently construct all ChangeSets in memory and then process each one.
We maintain pointers to all of the ChangeSets during this process, so the
overall memory footprint only increases due to the CommitVisitors.
See RepositoryMining.processRepo.

The "full-featured ChangeSet" PR increases the size of each ChangeSet,
and there is concern that this may lead to OOM issues on small machines.

Optimization:
Discard the pointer to each ChangeSet after we process it.
If a CommitVisitor does not save the Commits it visits, then
a ChangeSet can be GC'd after a worker processes it.

Longer-term suggestions:
1. If we stream ChangeSets instead of constructing them up front, this
   concern goes away entirely.
2. If we make the amount of data stored in a ChangeSet tunable, the
   old ChangeSet memory footprint can still be achieved.