bug: fix error in bookkeeping
Problem:
There is a bookkeeping assert for the threaded processing of ChangeSets.
The assert ensures that some thread attempted to work on every ChangeSet.

However, the per-thread counter is only incremented after attempting `processChangeSet`.
If `processChangeSet` throws an exception, the counter is not incremented.

The bookkeeping assert fails, but it should not because we still attempted
to work on the ChangeSet.

Solution:
Increment the counter before we `processChangeSet`.

Test:
1. This commit introduces a test for the case where a Visitor throws.
   It turns out that this case is innocuous because these are caught
	 by `RepoVisitor::visitCommit` and not thrown by `processChangeSet`.
2. You can, however, cause `processChangeSet` to throw by modifying
   the commit limits in `GitRepository`:

```java
private static final int MAX_SIZE_OF_A_DIFF = 100000;
private static final int DEFAULT_MAX_NUMBER_OF_FILES_IN_A_COMMIT = 5000;
```

This is concerning and should be addressed elsewhere.

Fix for: #118.